requirements.txt
absl-py==1.0.0
astunparse==1.6.3
bleach==4.1.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.8
colorama==0.4.4
cycler==0.11.0
docutils==0.18.1
flatbuffers==2.0
fonttools==4.28.2
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.42.0
h5py==3.6.0
idna==3.3
imageio==2.12.0
importlib-metadata==4.8.2
joblib==1.1.0
keras==2.7.0
Keras-Preprocessing==1.1.2
keyring==23.3.0
kiwisolver==1.3.2
libclang==12.0.0
Markdown==3.3.6
matplot==0.1.9
matplotlib==3.5.0
numpy==1.21.4
oauthlib==3.1.1
opencv-python==4.5.4.60
opt-einsum==3.3.0
packaging==21.3
Pillow==8.4.0
pkginfo==1.8.1
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.10.0
pyloco==0.0.139
pyparsing==3.0.6
python-dateutil==2.8.2
pywin32-ctypes==0.2.0
readme-renderer==30.0
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.5.0
rsa==4.8
scikit-learn==1.0.1
scipy==1.7.3
setuptools-scm==6.3.2
SimpleWebSocketServer==0.1.1
six==1.16.0
sklearn==0.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.22.0
termcolor==1.1.0
threadpoolctl==3.0.0
tomli==1.2.2
tqdm==4.62.3
twine==3.6.0
typing==3.7.4.3
typing_extensions==4.0.0
urllib3==1.26.7
ushlex==0.99.1
webencodings==0.5.1
websocket-client==1.2.1
Werkzeug==2.0.2
wrapt==1.13.3
zipp==3.6.0
